requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
face-recognition==1.3.0
facenet-pytorch==2.5.3
torch==2.1.0
torchvision==0.16.0
scipy==1.11.4
mediapipe==0.10.8

# Optional: dlib (difficult to install on Windows)
# If you want to use dlib, install it separately:
# Option 1: Using conda (recommended for Windows)
#   conda install -c conda-forge dlib
# Option 2: Download pre-built wheel from:
#   https://github.com/sachadee/Dlib/releases
# Option 3: Build from source (requires Visual Studio Build Tools)
# dlib==19.24.2

python-dotenv
python-dotenv
